Although some hppa models, such as the 715, will require specific keyboards, some others, like the 712, will run with a regular ps/2 keyboard from your average computer shop, and the hardware will convert the scancodes so that they match the braindead gsc usage.
Thus, it makes no point carrying a specific set of keyboard layouts for gsckbd,
when in fact it's (almost) the same as for pckbd, with scancodes shuffled.

So, add a simple scheme to build gsckbd tables from the pckbd tables. This
brings all the international layouts from pckbd, but jp and br have extra
keys for which I don't know the scancodes, so the generated gsckbd layouts
are incomplete.