Add power hook to scoop(4), replacing the Zaurus-specific scoop_* calls in pxa2x0_apm.c.
Set a flag to indicate that scoop is currently suspended.  Test this flag
in scoop_timeout() (and maybe later in other places) to avoid manipulating
scoop registers while the driver is suspended.  The scoop_checkdisk timeout
is intentionally not stopped, as it should be safe to have timeouts running
accross suspend/resume.

Put splhigh() in scoop_gpio_pin_write() to make changing the pin level an
atomic operation.

Result of prodding and diffs from Robert and Theo.