Add a TAILQ, aml_nodelist, of devices depending on the dock device to struct acpidock_softc and stuff devices into it during attach.
This list is not yet used, but I have code ready to handle the ACPI side
of this.  However, it still doesn't handle that there may in fact be
multiple dock devices in a machine (e.g. on pre-60 series ThinkPads), but
I need to figure out how to do so properly first.  In the mean time I want
this in the tree so I don't lose the code.