Add the long requested M_ZERO flag to malloc(9).
But the reason for this isn't some kind of "we can make it use the
pre-zeroed pages and zero the freelist in the idle loop and OMG I can
has optimisatiuns" which would require tons of infrastructure and make
everything slower.

The reason is that it shrinks other code. And that's good.

dlg@ ok, henning@ ok (before he read the diff)