pin 47 on ALC65{0,5,8} ac97 codecs can be used for either eapd or spdif-in.  usually the deault of spdif-in is used, but some ATI devices use pin 47 for eapd.
add a list of subsystem ids (from ALSA) for these devices, and set
bit 1 of register 7a to select the correct function of pin 47 for
these devices.

there was code here to do this prior to the ac97 update, but it
was accidentally removed.  rename the flag and bit to more accurately
reflect what this does and hopefully avoid future confusion.

problem noticed and fix tested/ok'd by matthieu@, thanks

also some discussion with and pointers from brad@, thanks