switch from using MCLGET to MCLGETI for clusters going onto the rx ring. this causes vic to start up with 4 mbufs on the ring, and the allocator will let it increase the number of rx mbufs as usage demands.
testing shows that i need only 20 to 30 mbufs on the rx ring to
cope with full speed io. we used to always put 100 on the ring, so
we're now saving 140k of kernerl virtual address space (70 *
MCLBYTES).

this relies on the previous commit that copes with failures of mbuf
allocation in the rx path.

similar changes must be made to all the other drivers. i'll happily review
diffs people send in. SUBTLE HINTS ARE COMING YOUR WAY.