Kill a bunch of DVACT_DEACTIVATE handlers whose only purpose were to set the 'dying' flag of a device.  Such handlers are useless now that usbd_detach() already set this flag.
Even if the purpose of this flag is questionnable on OpenBSD because
DVACT_DEACTIVATE acts as a pre-detach event, this is just a first small
step towards the simplification of the autoconf(9) device states.

This cleaning is now possible thanks to the work of pirofti@ to convert
all the USB drivers to properly use usbd_is_dying().

Discussed many times with deraadt@