KASSERTMSG(9): New kernel assertion with message
KASSERT() is annoying as it only prints the expression as a string.  If you
(developers) want to know a little more information, you have to do:

	#ifdef DIAGNOSTIC
		if (bad)
			panic(...);
	#endif

KASSERTMSG() replaces it into a single line:

		KASSERTMSG(!bad, ...);

Taken from NetBSD.

(There is a concern that KASSERT() messages are too long; consume more memory,
and not friendly for small monitors.  This have to be considered & revisited
later.)

"Like" from henning@
Man page review & advices from jmc@ and schwarze@