bring back r1.80:
pass the size of the softc to free.

this time we only dereference the cfattach structure if we actually
need it once the refcount has dropped to 0.

tested by krw@ who found the original panic r1.80 was backed out for.
putting this in again so i can find out when else will break.