Reduce the stride length of the tables by two and use a single page allocator for the 4K heap.
In this configuration a fullfeed BGP server for v4 and v6 consumes
10M more than with the radix tree.

This double the depth of the tree and makes the lookup slower.  But
the ratio speed/memory can be adjusted in the future, for now we are
interested in a lock-free route lookup.

Tested by and ok benno@