Prevent efifb(4) from attaching if we're the console.
Unfortunately, making this decision in radeondrm_attachhook() is too late
because at that point efifb(4) would have already been attached.  This means
that if we decide to bail in radeondrm_attachhook() we may end up without
a glass console.  That may happen for example if the firmware package
has not been installed.  I'm still looking for a solution for that problem.

ok jsg@