Properly support the mac context command in iwm. This was wrong in many ways. This commit includes style fixes as well as actual bug fixes.
Remove the global in_assoc flag from the softc. We can just pass a fixed
argument to functions to indicate whether we're already associated.

Remove in_tfsid from the softc as well. It was implicitly set to zero
and never changed. A corresponding macro already exists so use it instead.

Always copy the BSSID into the mac context command, as the Linux driver does.
This helps the firmware during association.

Set the firmware state to 'associated' once we've moved to RUN state
and our assoc ID is known. Earlier versions of iwm used to set the
firmware to 'associated' but this was disabled in r1.91 since it
didn't work correctly due to bugs which are now fixed.

Use live DTIM information obtained from the AP, rather than using the
default DTIM period of our wireless stack.

Check return values of functions called after moving to RUN state.

ok sthen@