Reflect interface priorities when inserting RTF_CONNECTED routes.
When multiple RTF_CLONING routes exist for a given subnet, bringing
one of the interfaces up/down could make impossible to insert new
ARP/NDP entries.

In this case the first RTF_CONNECTED route of the multipath list no
longer corresponded to the wired interface because it had the same
priority as the carp(4) route.

This is another regression of supporting multiple RTF_CONNECTED routes.

Fix a bug with a CARP setup reported by stsp@

ok stsp@