Implement Dynamic Profiling, a ddb(4) based & gprof compatible kernel profiling framework, for i386.
Code patching is used to enable probes when entering functions.  The
probes will call a mcount()-like function to match the behavior of a
GPROF kernel.

A new sysctl knob, ddb.profile, need to be set to 1 in securelevel 0
to be able to use this feature.

ok jasper@, guenther@, mlarkin@