Fix issues with iwm(4) firmware's retry rate table:
For a HT node, ni_txrate is always zero. We should be using ni_txmcs instead.
Simplify the if-else logic to make sure of that.

The mimo delimiter in the link quality command was never set.
I don't know how important this is. But Linux sets it, so why not.

Hardcode the lowest rate at the tail of the retry table.
While debugging the old code I have encountered retry tables filled
with only 'MCS 8' which is obviously not ideal.

While here, fix a misspelled function prototype.

Tested by kettenis@, mpi@, and Stefan Wollny
ok mpi@