tweak the spllower asm so it is more straightforward.
this properly identifies the registers used as input and output
operands to the code running in the trap handler, and passes them
to the asm statement as such. this means we dont have to do an extra
copy in the asm, or an extra clobber to keep the compiler away from
the registers. it also lets gcc set up and use the input register
nicely before it reaches the asm.

ok kettenis@