Tighten up FQ-CoDel vs HFSC checks
PFQS_FLOWQUEUE is about to become a flag that HFSC classes may specify
as well; thus a better way of telling FQ-CoDel and HFSC apart needs to
be found.  At the moment its derived from the queue specification.

Discussed with and OK henning@ at d2k17 as a part of a larger diff.