Add a PCI attachment driver for bwfm(4).  It's not finished, but it's already past the point where development can occur out of the tree. With this I can successfully scan for access points and tell the chip to attach to an SSID.  RX path should work as well, but since I forgot to bring the antenna with me to my parents, the reception is a bit horrible in the metal enclosure.
There are a few reasons this driver is rather big.  First we set up the
ARM Cores, uploading the firmware and kicking it off.  Then we need to
read all needed information from the registers.  Once that is done we
have to set up countless buffers.  There are 2 TX rings and 3 RX rings,
plus N TX rings for the actual data that is yet to be implemented.

Merry Christmas!

ok kettenis@