vmm(4): fix lockups on SVM hosts when booting Linux guests
The previous version of the code used "disable_intr" as a protection
around the FPU state saving routine, which left host RFLAGS.IF=0 when
entering the guest. If the guest ended up in a non-exitable condition,
the machine could hang due to missed TLB flush IPIs (and eventual
kernel lock contention and starvation).

tested by brynet@

ok guenther@