pgt(4): do not tsleep in the scan ioctl.
The only action taken during the scan ioctl was to sleep. This does not make
sense as the scanning process happens in the state machine behind the scenes.

Initialization indeed performs newstate(SCAN), but that is done no matter
if the user waits for the full results or not. Instead, this change unifies
the (new) user experience when performing ifconfig scan and shows the
nodes list produced by the latest background scan.

This was not tested on real hardware.
Please message me directly in case this broke your card.

OK stsp@, deraadt@