Harmonize spacing after ellipses in displayed messages.
We were using spacing after ellipses in an inconsistent way in the
installer. Standardize on using "... " everywhere and take into account
the cursor position while we are waiting for the task to complete: the
cursor is now always positioned after the last dot, and the space is
added when displaying completion confirmation.

While there, also take cursor position into account in vfs_shutdown(),
and remove the extra leading space before ticks in dhclient.

OK deraadt@