Set RUNNING flag when RF switch is enabled
Toggling the hardware kill switch shuts radio power on/off accordingly.
On UP and RUNNING interfaces disabling removes the RUNNING flag,
but enabling back again would leave the flag unset.

Get the interface running when switching it back on iff it's up
to ensure seemless operation without the need for `ifconfig iwn0 up'.

For this, iwn_init_task() will do the right thing so schedule it when
receiving the appropiate interrupt.

Tested on my X230 with
    "Intel Centrino Advanced-N 6205" rev 0x34: msi, MIMO 2T2R, MoW

OK stsp who also tested it with
    "Intel Centrino Advanced-N 6200" rev 0x35: msi, MIMO 2T2R, MoW

iwm(4) already behaves this way.