vmm: better handling of two SMM related MSRs
We currently ignore MSR_SMBASE and MSR_SMM_MONITOR_CTL, but the SDM says
accessing the former for read and latter for write while not in SMM mode
should produce a #GP. This change detects those operations and injects
a #GP as the documentation says. The previous behaviour was harmless, just
not correct.

ok pd