reintroduce ifiq_input counting backpressure
instead of counting the number of packets on an ifiq, count the
number of times a nic has tried to queue packets before the stack
processes them.

this new semantic interacted badly with virtual interfaces like
vlan and trunk, but these interfaces have been tweaked to call
if_vinput instead of if_input so their packets are processed directly
because theyre already running inside the stack.

im putting this in so we can see what the effect is. if it goes
badly i'll back it out again.

ok cheloha@ proctor@ visa@