Use timeout_add_msec(9)
As per termios(4), "VTIME is a timer of 0.1 second granularity", so
convert it to milliseconds by reducing by hz and multiplying with 1000.

Furthermore, the specification (and our implementation) define members
of the c_cc arry to be of type unsigned char, so both the previous as
well as the now used arithmetic operations are guaranteed to not
overflow.

Since the timeout_add(9) API takes an int argument, the previous long
type would always be demoted anyway, so change it to int directly.

With this and hz gone, remove the obselete comment.

While here, use more mnemonic variable names.

Feedback and OK mpi