pay some lip service to TPMR compliance according to 802.1Q-2018
the spec says we should filter packets destined to a list of ethernet
addresses. im currently interpreting "filter" as meaning dropping,
which this diff does.

however, one of the addresses to filter is the one lacp uses by
default and not a lot of lacp implementations (read switches) support
the configuration of a different address. i still need lacp to go
over tpmr, and because i can't change the address, this diff also
has a way to configure tpmr to still allow the packets through.