Add RTP_PROPOSAL_SOLICIT to allow unwind(8) to solicit DNS proposals from slaacd and dhclient when it starts.
Discussed with deraadt who notes that it's a bit odd to have this as a
route priority. One idea is to have this as a dedicated route message
and not a priority.

But we want to move this forward and learn how it can be used so we
are going with this for now.

OK deraadt