Do not delete an existing RTF_CACHED entry with the same destination address as the one trying to be inserted.
Such entry must stay in the table as long as its parent route exist.  If
a code path tries to re-insert a route with the same destination address
on the same interface it is a bug.

Avoid the "route contains no arp information" problem reported by sthen@
and Laurent Salle.

ok claudio@