Don't enable any BT modules in the iwx(4) BT_COEX command.
The iwx(4) driver intends to disable bluetooth. The command we sent to the
firmware says "wifi only", but also enabled a bluetooth module. This seems
to work as intended but is ambiguous. Linux also leaves all modules disabled
in wifi-only mode. So do not enable any BT modules in the BT_COEX command.

ok mpi@