Make iwx(4) update the Tx ring byte count table.
With iwn(4) and iwm(4) this was only required for Tx aggregation, which
we don't use yet in iwx(4). But it looks like iwx(4) firmware will run
into fatal errors if we neglect to update this table. The Linux iwlwifi
driver always updates this table so firmware could be relying on it.

Tested by jcs@, sven falempin, and myself.