move vlan_input into ether_input, instead of via an input handler.
this means there's a consistent order of processing of service
delimited (vlan and svlan) packets and bridging of packets. vlan
and svlan get to look at a packet first. it's only if they decline
a packet that a bridge can handle it. this allows operators to slice
vlans out for processing separate to the "native" vlan handling if
they want.

while here, this fixes up a bug in vlan_input if m_pullup needed
to prepend an mbuf.

this has been in snaps as part of a larger diff for over a week.