setitimer(2): write new timer value in one place
Rearrange the critical section in setitimer(2) to match that of
getitimer(2).  This will make it easier to merge the two critical
sections in a subsequent diff.

In particular, we want to write the new timer value in *one* place in
the code, regardless of which timer we're setting.

ok millert@