Change kqueue_scan() to keep track of collected events in the given context.
It is now possible to call the function multiple times to collect events.
For that, the end marker has to be preserved between calls because otherwise
the scan might collect an event more than once.  If a collected event gets
reactivated during scanning, it will be added at the tail of the queue,
out of reach because of the end marker.

This is required to implement select(2) and poll(2) on top of kqueue_scan().

Done & originally committed by visa@ in r1.143, in snap for more than 2 weeks.

ok visa@, anton@