tht(4): more tsleep(9) -> tsleep_nsec(9)
The first wait-loop in tht_fw_load() can sleep for at least 10ms per
iteration instead of up to 1 tick per iteration with no ill effects.
The worst-case scenario is that we sleep a little longer than we do
now while we wait.

The second wait-loop can be simplified to sleep for at least 10ms per
iteration for up to 2000ms instead of using a timeout.  This method is
less precise and less efficient but is far simpler to read and
understand than using a timeout.  We can then remove all the
timeout-related stuff from if_tht.c.

Discussed with mpi@.

ok dlg@