Fix the return value for the FUTEX_WAIT/FUTEX_WAIT_PRIVATE futex(2) operation.  System calls should return -1 and set errno when they fail. They should not return an errno value directly.  This matches how the Linux version of futex(2) behaves and what Mesa expects.  This fixes a bug in Mesa where a timeout wouldn't be reported properly.
Technically this is an ABI break.  But libc and libpthread were changed
to be compatible with both the old and new ABI, and code outside of base
almost certainly expects Linux compatible behaviour.  If you have not
rebuilt libc and the last few days, upgrade using a snap.

Mesa issue discovered by jsg@
ok mpi@, deraadt@