In iwm(4) and iwx(4), prevent attemps to transition towards the same state in cases where this would result in a redundant or illegal state transition.
jmc@ observed ASSOC -> ASSOC transitions which would result in a hang.
Such transitions are invalid and never intentionally triggered by net80211.
They imply a race between the Rx interrupt handler and the newstate task.

Tested by jmc@ on AX200 for a week and several known issues seem to be fixed.