Remove unused AUDIO_PROP_{MMAP,INDEPENDENT}
AUDIO_PROP_FULLDUPLEX is the only audio(9) in use, the other two died with
	commit 6eb34baff6cbf8b5729e60178074cd13c4817e0a
	Author: ratchov <ratchov@openbsd.org>
	Date:   Thu Jun 25 06:43:45 2015 +0000

	Reimplement the audio driver in a simpler way, removing unused/unusable
	functionality. Same API and ABI except for the removed bits and no
	behaviour change for programs using libsndio. With help from armani@
	and mpi@, thanks.

but remained defined and set in drivers.

and the following merely moved them when they were dead code already:

	commit e7f6b574696d210437f446e2a4c3d82741bcdb33
	Author: ratchov <ratchov@openbsd.org>
	Date:   Wed Aug 31 07:22:43 2016 +0000

	    Delete unused ioctls and associated macros. Move macros that are still
	    used internally by low-level drivers from sys/audioio.h to
	    dev/audio_if.h instead of deleting them.

None of this is used in base or ports;  codesearch.debian.net only shows
AUDIO_PROP_{CAPTURE,PLAYBACK} in firefox-esr, mozjs and cubeb.

ratchov points out that audio_if.h and audioio.h are private interfaces
and the codesearch shows SunOS and NetBSD bits (#ifdef'd out on OpenBSD).

OK ratchov